refactor: split wagmi provider config into its own file and component
